requirements.txt
sqlmodel
psycopg[binary,pool]
fastapi[standard]
uvicorn[standard]
gunicorn 
orjson
python-dotenv